+++ src/main/java/com/project/Teaming/TeamingApplication.java
@@ -11,7 +11,6 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @SpringBootApplication
 @EnableJpaAuditing
 @EnableScheduling
-@EnableAsync
 @OpenAPIDefinition(servers = {@Server(url = "https://myspringserver.shop", description = "teaming server")})
 public class TeamingApplication {
