+++ src/main/java/com/project/Teaming/domain/mentoring/repository/ParticipationRepositoryCustomImpl.java
@@ -46,7 +46,6 @@ public class ParticipationRepositoryCustomImpl implements ParticipationRepositor
                         u.name,
                         mp.role,
                         mp.participationStatus,
-                        mp.isDeleted,
                         new CaseBuilder()
                                 .when(mt.status.eq(teamStatus)
                                         .and(r.id.isNotNull()))
@@ -112,7 +111,6 @@ public class ParticipationRepositoryCustomImpl implements ParticipationRepositor
                         .join(mp.mentoringTeam, mt).fetchJoin()
                         .where(
                                 mt.id.eq(teamId),
-                                mp.isDeleted.isFalse(),
                                 mp.participationStatus.eq(participationStatus),
                                 mp.authority.eq(authority)
                         )
@@ -133,7 +131,6 @@ public class ParticipationRepositoryCustomImpl implements ParticipationRepositor
                 .where(
                         mp.user.eq(user),
                         mp.participationStatus.eq(status),
-                        mp.isDeleted.isFalse(),
                         mt.flag.eq(Status.FALSE)
                 )
                 .distinct()
